Fixed compiler warning in MprSpeakerSelector.cpp
git-svn-id: https://sipxsvn.sipez.com/rep/sipX/main@12388 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaLib/src/mp/MprSpeakerSelector.cpp
@@ -1,6 +1,5 @@
 //
-// Copyright (C) 2008 SIPez LLC. 
-// Licensed to SIPfoundry under a Contributor Agreement. 
+// Copyright (C) 2008-2017 SIPez LLC. All rights reserved.
 //
 // Copyright (C) 2008 SIPfoundry Inc.
 // Licensed by SIPfoundry under the LGPL license.
@@ -614,7 +613,7 @@ void MprSpeakerSelector::updateMapping(MprSpeakerSelector::RankIndexPair *topRan
          }
          else if (mOutToInMap[mInToOutMap[speaker]] != speaker)
          {
-            if (mChangeMapTime[mInToOutMap[speaker]] != mMapTime)
+            if (mChangeMapTime[mInToOutMap[speaker]] != (int)mMapTime)
             {
                // This input may know its place and may reside on it.
                // Its competitor has not been active for a while.
